Supporting Legacy Serialization Format When Receiving
The legacy serialization format of a tuple array
was causing an issue with the lua command when checking
if we should delete the unique id out of the set.

We'd specifically get the following error:

@user_script:26: @user_script: 26: Lua redis() command arguments must be strings or integers

which stemmed from the fact that when we tested if uniqueId field didn't equal cjson.nil,
we essentially allowed for any non-nil type to be executed, but in reality we should
only allow strings to be passed through. So we made that update.

Signed-off-by: RJ Garcia <rj@stadiumgoods.com>